// core/RUNTIME/TEMPORAL/types.ts
export enum TemporalStatus {
  ACTUAL = 'ACTUAL',
  MEMORY = 'MEMORY',
  REPORT = 'REPORT',
  IMAGINED = 'IMAGINED',
  HYPOTHETICAL = 'HYPOTHETICAL',
  PLAN = 'PLAN',
  PREDICTION = 'PREDICTION',
  POSSIBILITY = 'POSSIBILITY',
  UNKNOWN = 'UNKNOWN'
}